requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai>=1.40.0
python-dotenv==1.0.0
httpx>=0.27.0
langfuse>=2.0.0
python-docx>=1.1.0
pymongo>=4.6.0

# Voice processing dependencies
faster-whisper>=0.10.0
pyttsx3>=2.90
pydub>=0.25.1

# HealthBench Evaluation Dependencies
blobfile>=2.0.0
chromadb>=0.4.0

